requirements.txt
Flask
python-binance
pandas
ta
python-dotenv
